_Marketing/gtm-mcp/gtm-mcp-main/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtm-mcp"
version = "0.1.0"
description = "GTM MCP — B2B lead generation tools for AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.118.0",
    "google-auth>=2.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
gtm-mcp = "gtm_mcp.server:main"
